+++ frontend/components/Footer.tsx
@@ -1,20 +1,6 @@
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/utc";
-import utc from "dayjs/plugin/relativeTime";
 import { Status } from "../lib/status";
 
-type FooterProps = {
-  status: Status;
-};
-
-const Footer = ({ status }: FooterProps) => {
-  dayjs.extend(relativeTime);
-  dayjs.extend(utc);
-
-  const updatedOnUtc = dayjs.utc(status.updated_on);
-  const utcNow = new Date().toUTCString();
-  const lastUpdated = updatedOnUtc.from(utcNow);
-
+const Footer = () => {
   return (
     <div className="text-center text-gray-900 dark:text-gray-100">
       <p className="font-bold hover:underline decoration-yellow-400 decoration-2 mb-8">
@@ -23,7 +9,6 @@ const Footer = ({ status }: FooterProps) => {
           Or better exchange rates? Let us know!
         </a>
       </p>
-      <p className="text-sm">Last updated {lastUpdated}</p>
       <p className="text-xs mt-8 text-gray-500">
         <a
           target="_blank"
